+++ R/summary.spwb.R
@@ -1,3 +1,33 @@
+.mergeVectorOfMatrices<-function(vec){
+  out<- NULL
+  for(i in 1:length(vec)) {
+    out_i <- vec[[i]]
+    if(is.null(out)) out = out_i
+    else {
+      cno <- colnames(out)
+      cno_i <- colnames(out_i)
+      if(length(cno)==length(cno_i) && sum(cno==cno_i)==length(cno)) {
+        out<-rbind(out, out_i)
+      } else {
+        cn_all = unique(c(cno, cno_i))
+        cn_old = cno[which(!(cno %in% cno_i))]
+        cn_new = cno_i[which(!(cno_i %in% cno))]
+        if(length(cn_new)>0) {
+          out <- as.data.frame(out)
+          for(n in cn_new) out[[n]] = NA
+          out <- as.matrix(out[,cn_all])
+        }
+        if(length(cn_old)>0) {
+          out_i <- as.data.frame(out_i)
+          for(n in cn_old) out_i[[n]] = NA
+          out_i <- as.matrix(out_i[,cn_all])
+        }
+        out<-rbind(out, out_i)
+      }
+    }
+  }
+  return(out)
+}
 .summarysim<-function(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...){  
   dates = as.Date(rownames(object$WaterBalance))
   ndaysTotal = length(dates)
@@ -80,34 +110,10 @@ summary.growth<-function(object, freq="years", output="WaterBalance", FUN=sum, b
 }
 
 summary.fordyn<-function(object, freq="years", output="WaterBalance", FUN=sum, bySpecies = FALSE, ...){
-  out<-NULL
+  vec<-vector("list", length(object$GrowthResults))
   for(i in 1:length(object$GrowthResults)) {
-    out_i <- .summarysim(object = object$GrowthResults[[i]], 
-                     freq = freq, output = output, FUN = FUN, bySpecies = bySpecies,
-                     )
-    if(is.null(out)) out = out_i
-    else {
-      cno <- colnames(out)
-      cno_i <- colnames(out_i)
-      if(length(cno)==length(cno_i) && sum(cno==cno_i)==length(cno)) {
-        out<-rbind(out, out_i)
-      } else {
-        cn_all = unique(c(cno, cno_i))
-        cn_old = cno[which(!(cno %in% cno_i))]
-        cn_new = cno_i[which(!(cno_i %in% cno))]
-        if(length(cn_new)>0) {
-          out <- as.data.frame(out)
-          for(n in cn_new) out[[n]] = NA
-          out <- as.matrix(out[,cn_all])
-        }
-        if(length(cn_old)>0) {
-          out_i <- as.data.frame(out_i)
-          for(n in cn_old) out_i[[n]] = NA
-          out_i <- as.matrix(out_i[,cn_all])
-        }
-        out<-rbind(out, out_i)
-      }
-    }
+    vec[[i]] <- .summarysim(object = object$GrowthResults[[i]], 
+                           freq = freq, output = output, FUN = FUN, bySpecies = bySpecies, ...)
   }
-  return(out)
+  return(.mergeVectorOfMatrices(vec))
 }